requirements.txt
fastapi
uvicorn
aiohttp
beautifulsoup4
python-dotenv
scikit-learn
nltk
numpy
bs4
langchain
langchain-google-genai
google-generativeai 